* Updated locatortest to new network stack standards
* Changed uart to handle baud rate values larger than 16 bits

* Restored master-receiver mode of TWI 


git-svn-id: https://svn.circlemud.org/svn/rocket/trunk@630 46170a15-b891-4bf7-af97-fd374ce181e0

+++ src/include/hal.h
@@ -77,7 +77,7 @@ typedef struct s_UartHandler {
 	hal_uart_send_next_fp *send;	// runs in interrupt context
 } UartHandler;
 
-void hal_uart_init(UartHandler *handler, uint16_t baud, r_bool stop2);
+void hal_uart_init(UartHandler *handler, uint32_t baud, r_bool stop2);
 void hal_uart_start_send(void);
 
 /////////////// TWI ///////////////////////////////////////////////
